allow host to add attendees
